server/auth/requirements.txt
aiofiles==0.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
starlette==0.13.6
toml==0.10.2
urllib3==1.26.3
uvicorn==0.13.3
uvloop==0.14.0
websockets==8.1
